+++ src/IRReceive.h
@@ -1,0 +1,295 @@
+/**
+ * @file IRReceive.h
+ * @brief Contains all declarations required for the internal functions.
+ *
+ * This file is part of Arduino-IRremote https://github.com/Arduino-IRremote/Arduino-IRremote.
+ *
+ *
+ ************************************************************************************
+ * MIT License
+ *
+ * Copyright (c) 2015-2021 Ken Shirriff http://www.righto.com, Rafi Khan, Armin Joachimsmeyer
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is furnished
+ * to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
+ * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE
+ * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *
+ ************************************************************************************
+ */
+
+#ifndef IRReceive_h
+#define IRReceive_h
+
+#include "IRremoteIntDef.h"
+
+/****************************************************
+ *                     RECEIVING
+ ****************************************************/
+/*
+ * Activating this saves 60 bytes program space and 14 bytes RAM
+ */
+//#define NO_LEGACY_COMPATIBILITY
+#if !defined(NO_LEGACY_COMPATIBILITY)
+/**
+ * Results returned from old decoders !!!deprecated!!!
+ */
+struct decode_results {
+    decode_type_t decode_type;  // deprecated, moved to decodedIRData.protocol ///< UNKNOWN, NEC, SONY, RC5, ...
+    uint16_t address;           ///< Used by Panasonic & Sharp [16-bits]
+    uint32_t value;             // deprecated, moved to decodedIRData.decodedRawData ///< Decoded value / command [max 32-bits]
+    uint8_t bits;               // deprecated, moved to decodedIRData.numberOfBits ///< Number of bits in decoded value
+    uint16_t magnitude;         // deprecated, moved to decodedIRData.extra ///< Used by MagiQuest [16-bits]
+    bool isRepeat;              // deprecated, moved to decodedIRData.flags ///< True if repeat of value is detected
+
+// next 3 values are copies of irparams values - see IRremoteint.h
+    uint16_t *rawbuf;           // deprecated, moved to decodedIRData.rawDataPtr->rawbuf ///< Raw intervals in 50uS ticks
+    uint16_t rawlen;            // deprecated, moved to decodedIRData.rawDataPtr->rawlen ///< Number of records in rawbuf
+    bool overflow;              // deprecated, moved to decodedIRData.flags ///< true if IR raw code too long
+};
+#endif
+
+/*
+ * Definitions for member IRData.flags
+ */
+#define IRDATA_FLAGS_EMPTY              0x00
+#define IRDATA_FLAGS_IS_REPEAT          0x01
+#define IRDATA_FLAGS_IS_AUTO_REPEAT     0x02
+#define IRDATA_FLAGS_PARITY_FAILED      0x04 ///< the current (autorepeat) frame violated parity check
+#define IRDATA_TOGGLE_BIT_MASK          0x08
+#define IRDATA_FLAGS_EXTRA_INFO         0x10 ///< there is unexpected extra info not contained in address and data (e.g. Kaseikyo unknown vendor ID)
+#define IRDATA_FLAGS_WAS_OVERFLOW       0x40 ///< irparams.rawlen is 0 in this case to avoid endless OverflowFlag
+#define IRDATA_FLAGS_IS_LSB_FIRST       0x00
+#define IRDATA_FLAGS_IS_MSB_FIRST       0x80 ///< Just for info. Value is simply determined by the protocol
+
+/**
+ * Data structure for the user application, available as decodedIRData.
+ * Filled by decoders and read by print functions or user application.
+ */
+struct IRData {
+    decode_type_t protocol;     ///< UNKNOWN, NEC, SONY, RC5, ...
+    uint16_t address;           ///< Decoded address
+    uint16_t command;           ///< Decoded command
+    uint16_t extra;             ///< Used by MagiQuest and for Kaseikyo unknown vendor ID.  Ticks used for decoding Distance protocol.
+    uint16_t numberOfBits;      ///< Number of bits received for data (address + command + parity) - to determine protocol length if different length are possible.
+    uint8_t flags;              ///< See IRDATA_FLAGS_* definitions above
+    uint32_t decodedRawData;    ///< Up to 32 bit decoded raw data, used for sendRaw functions.
+    irparams_struct *rawDataPtr; ///< Pointer of the raw timing data to be decoded. Mainly the data buffer filled by receiving ISR.
+};
+
+/**
+ * Just for better readability of code
+ */
+#define USE_DEFAULT_FEEDBACK_LED_PIN 0
+
+/**
+ * Main class for receiving IR signals
+ */
+class IRrecv {
+public:
+
+    IRrecv();
+    IRrecv(uint8_t aReceivePin);
+    IRrecv(uint8_t aReceivePin, uint8_t aFeedbackLEDPin);
+    void setReceivePin(uint8_t aReceivePinNumber);
+
+    void enableIRIn();
+    void disableIRIn();
+
+    /*
+     * Stream like API
+     */
+    void begin(uint8_t aReceivePin, bool aEnableLEDFeedback = false, uint8_t aFeedbackLEDPin = USE_DEFAULT_FEEDBACK_LED_PIN);
+    void start(); // alias for enableIRIn
+    void start(uint32_t aMicrosecondsToAddToGapCounter);
+    bool available();
+    IRData* read(); // returns decoded data
+    // write is a method of class IRsend below
+    // size_t write(IRData *aIRSendData, uint_fast8_t aNumberOfRepeats = NO_REPEATS);
+    void stop(); // alias for disableIRIn
+    void end();
+
+    bool isIdle();
+
+    /*
+     * The main functions
+     */
+    bool decode();  // Check if available and try to decode
+    void resume();  // Enable receiving of the next value
+
+    /*
+     * Useful info and print functions
+     */
+    void printIRResultShort(Print *aSerial);
+    void printIRResultMinimal(Print *aSerial);
+    void printIRResultRawFormatted(Print *aSerial, bool aOutputMicrosecondsInsteadOfTicks = true);
+    void printIRResultAsCVariables(Print *aSerial);
+
+    void compensateAndPrintIRResultAsCArray(Print *aSerial, bool aOutputMicrosecondsInsteadOfTicks = true);
+    void compensateAndPrintIRResultAsPronto(Print *aSerial, unsigned int frequency = 38000U);
+
+    /*
+     * Store the data for further processing
+     */
+    void compensateAndStoreIRResultInArray(uint8_t *aArrayPtr);
+    size_t compensateAndStorePronto(String *aString, unsigned int frequency = 38000U);
+
+    /*
+     * The main decoding functions used by the individual decoders
+     */
+    bool decodePulseDistanceData(uint8_t aNumberOfBits, uint8_t aStartOffset, uint16_t aBitMarkMicros, uint16_t aOneSpaceMicros,
+            uint16_t aZeroSpaceMicros, bool aMSBfirst);
+
+    bool decodePulseWidthData(uint8_t aNumberOfBits, uint8_t aStartOffset, uint16_t aOneMarkMicros, uint16_t aZeroMarkMicros,
+            uint16_t aBitSpaceMicros, bool aMSBfirst);
+
+    bool decodeBiPhaseData(uint_fast8_t aNumberOfBits, uint_fast8_t aStartOffset, uint_fast8_t aStartClockCount,
+            uint_fast8_t aValueOfSpaceToMarkTransition, uint16_t aBiphaseTimeUnit);
+
+    void initBiphaselevel(uint8_t aRCDecodeRawbuffOffset, uint16_t aBiphaseTimeUnit);
+    uint8_t getBiphaselevel();
+
+    /*
+     * All standard (decode address + command) protocol decoders
+     */
+    bool decodeBoseWave();
+    bool decodeDenon();
+    bool decodeJVC();
+    bool decodeKaseikyo();
+    bool decodeLegoPowerFunctions();
+    bool decodeLG();
+    bool decodeMagiQuest(); // not completely standard
+    bool decodeNEC();
+    bool decodeRC5();
+    bool decodeRC6();
+    bool decodeSamsung();
+    bool decodeSharp(); // redirected to decodeDenon()
+    bool decodeSony();
+
+    bool decodeDistance();
+
+    bool decodeHash();
+
+    // Template function :-)
+    bool decodeShuzu();
+
+    /*
+     * Old functions
+     */
+#if !defined(NO_LEGACY_COMPATIBILITY)
+    bool decodeDenonOld(decode_results *aResults);
+    bool decodeJVCMSB(decode_results *aResults);
+    bool decodeLGMSB(decode_results *aResults);
+    bool decodeNECMSB(decode_results *aResults);
+    bool decodePanasonicMSB(decode_results *aResults);
+    bool decodeSonyMSB(decode_results *aResults);
+    bool decodeSAMSUNG(decode_results *aResults);
+    bool decodeHashOld(decode_results *aResults);
+
+    bool decode(
+            decode_results *aResults)
+                    __attribute__ ((deprecated ("Please use IrReceiver.decode() without a parameter and IrReceiver.decodedIRData.<fieldname> ."))); // deprecated
+#endif
+    bool decodeWhynter();
+
+    // for backward compatibility. Now in IRFeedbackLED.hpp
+    void blink13(bool aEnableLEDFeedback)
+            __attribute__ ((deprecated ("Please use setLEDFeedback() or enableLEDFeedback() / disableLEDFeedback()."))); // deprecated
+
+    /*
+     * Internal functions
+     */
+    void initDecodedIRData();
+    uint8_t compare(unsigned int oldval, unsigned int newval);
+
+    IRData decodedIRData;       // New: decoded IR data for the application
+
+    // Last decoded IR data for repeat detection
+    decode_type_t lastDecodedProtocol;
+    uint32_t lastDecodedAddress;
+    uint32_t lastDecodedCommand;
+
+    uint8_t repeatCount;        // Used e.g. for Denon decode for autorepeat decoding.
+};
+
+extern uint8_t sBiphaseDecodeRawbuffOffset; //
+
+/*
+ * Mark & Space matching functions
+ */
+bool matchTicks(uint16_t aMeasuredTicks, uint16_t aMatchValueMicros);
+bool matchMark(uint16_t aMeasuredTicks, uint16_t aMatchValueMicros);
+bool matchSpace(uint16_t aMeasuredTicks, uint16_t aMatchValueMicros);
+
+/*
+ * Old function names
+ */
+bool MATCH(uint16_t measured, uint16_t desired);
+bool MATCH_MARK(uint16_t measured_ticks, uint16_t desired_us);
+bool MATCH_SPACE(uint16_t measured_ticks, uint16_t desired_us);
+
+int getMarkExcessMicros();
+
+void printIRResultShort(Print *aSerial, IRData *aIRDataPtr, uint16_t aLeadingSpaceDuration = 0);
+
+/****************************************************
+ * Feedback LED related functions
+ ****************************************************/
+void setFeedbackLED(bool aSwitchLedOn);
+void setLEDFeedback(uint8_t aFeedbackLEDPin, bool aEnableLEDFeedback); // if aFeedbackLEDPin == 0, then take board BLINKLED_ON() and BLINKLED_OFF() functions
+void setLEDFeedback(bool aEnableLEDFeedback); // Direct replacement for blink13()
+void enableLEDFeedback();
+void disableLEDFeedback();
+
+void setBlinkPin(uint8_t aFeedbackLEDPin) __attribute__ ((deprecated ("Please use setLEDFeedback()."))); // deprecated
+
+/**
+ * microseconds per clock interrupt tick
+ */
+#if ! defined(MICROS_PER_TICK)
+#define MICROS_PER_TICK    50
+#endif
+
+/*
+ * Pulse parms are ((X*50)-100) for the Mark and ((X*50)+100) for the Space.
+ * First MARK is the one after the long gap
+ * Pulse parameters in uSec
+ */
+/** Relative tolerance (in percent) for some comparisons on measured data. */
+#define TOLERANCE       25
+
+/** Lower tolerance for comparison of measured data */
+//#define LTOL            (1.0 - (TOLERANCE/100.))
+#define LTOL            (100 - TOLERANCE)
+/** Upper tolerance for comparison of measured data */
+//#define UTOL            (1.0 + (TOLERANCE/100.))
+#define UTOL            (100 + TOLERANCE)
+
+//#define TICKS_LOW(us)   ((int)(((us)*LTOL/MICROS_PER_TICK)))
+//#define TICKS_HIGH(us)  ((int)(((us)*UTOL/MICROS_PER_TICK + 1)))
+#if MICROS_PER_TICK == 50 && TOLERANCE == 25           // Defaults
+#define TICKS_LOW(us)   ((us)/67 )     // (us) / ((MICROS_PER_TICK:50 / LTOL:75 ) * 100)
+#define TICKS_HIGH(us)  ((us)/40 + 1)  // (us) / ((MICROS_PER_TICK:50 / UTOL:125) * 100) + 1
+#else
+    #define TICKS_LOW(us)   ((uint16_t) ((long) (us) * LTOL / (MICROS_PER_TICK * 100) ))
+    #define TICKS_HIGH(us)  ((uint16_t) ((long) (us) * UTOL / (MICROS_PER_TICK * 100) + 1))
+#endif
+
+/*
+ * The receiver instance
+ */
+extern IRrecv IrReceiver;
+
+#endif
